+++ src/main/java/school/faang/user_service/filters/user/UserAboutFilter.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.service.user.filter;
+package school.faang.user_service.filters.user;
 
 import org.springframework.stereotype.Component;
 import school.faang.user_service.dto.filter.UserFilterDto;
@@ -7,15 +7,15 @@ import school.faang.user_service.entity.User;
 import java.util.stream.Stream;
 
 @Component
-public class UserCountryFilter implements UserFilter {
+public class UserAboutFilter implements UserFilter{
     @Override
     public boolean isApplicable(UserFilterDto filters) {
-        return filters.getCountryPattern() != null;
+        return filters.getAboutPattern() != null;
     }
 
     @Override
     public Stream<User> apply(Stream<User> users, UserFilterDto filters) {
-        return users.filter(user -> user.getCountry().getTitle().toLowerCase()
-                .contains(filters.getCountryPattern().toLowerCase()));
+        return users.filter(user -> user.getAboutMe().toLowerCase()
+                .contains(filters.getAboutPattern().toLowerCase()));
     }
 }
