+++ src/main/java/school/faang/user_service/filters/user/UserCityFilter.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.service.user.filter;
+package school.faang.user_service.filters.user;
 
 import org.springframework.stereotype.Component;
 import school.faang.user_service.dto.filter.UserFilterDto;
@@ -7,14 +7,15 @@ import school.faang.user_service.entity.User;
 import java.util.stream.Stream;
 
 @Component
-public class UserNameFilter implements UserFilter {
+public class UserCityFilter implements UserFilter {
     @Override
     public boolean isApplicable(UserFilterDto filters) {
-        return filters.getNamePattern() != null;
+        return filters.getCityPattern() != null;
     }
 
     @Override
     public Stream<User> apply(Stream<User> users, UserFilterDto filters) {
-        return users.filter(user -> user.getUsername().contains(filters.getNamePattern()));
+        return users.filter(user -> user.getCity().toLowerCase()
+                .contains(filters.getCityPattern().toLowerCase()));
     }
 }
