+++ src/main/java/school/faang/user_service/filters/user/UserCountryFilter.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.service.user.filter;
+package school.faang.user_service.filters.user;
 
 import org.springframework.stereotype.Component;
 import school.faang.user_service.dto.filter.UserFilterDto;
@@ -7,14 +7,15 @@ import school.faang.user_service.entity.User;
 import java.util.stream.Stream;
 
 @Component
-public class UserPhoneFilter implements UserFilter {
+public class UserCountryFilter implements UserFilter {
     @Override
     public boolean isApplicable(UserFilterDto filters) {
-        return filters.getPhonePattern() != null;
+        return filters.getCountryPattern() != null;
     }
 
     @Override
     public Stream<User> apply(Stream<User> users, UserFilterDto filters) {
-        return users.filter(user -> user.getPhone().contains(filters.getPhonePattern()));
+        return users.filter(user -> user.getCountry().getTitle().toLowerCase()
+                .contains(filters.getCountryPattern().toLowerCase()));
     }
 }
