+++ src/main/java/school/faang/user_service/filters/user/UserPhoneFilter.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.service.user.filter;
+package school.faang.user_service.filters.user;
 
 import org.springframework.stereotype.Component;
 import school.faang.user_service.dto.filter.UserFilterDto;
@@ -7,14 +7,14 @@ import school.faang.user_service.entity.User;
 import java.util.stream.Stream;
 
 @Component
-public class UserEmailFilter implements UserFilter {
+public class UserPhoneFilter implements UserFilter {
     @Override
     public boolean isApplicable(UserFilterDto filters) {
-        return filters.getEmailPattern() != null;
+        return filters.getPhonePattern() != null;
     }
 
     @Override
     public Stream<User> apply(Stream<User> users, UserFilterDto filters) {
-        return users.filter(user -> user.getEmail().contains(filters.getEmailPattern()));
+        return users.filter(user -> user.getPhone().contains(filters.getPhonePattern()));
     }
 }
