+++ src/test/java/school/faang/user_service/service/event/EventServiceTest.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.service.service.event;
+package school.faang.user_service.service.event;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
