+++ src/test/java/school/faang/user_service/service/mentorship/MentorshipServiceTest.java
@@ -11,7 +11,7 @@ import school.faang.user_service.entity.User;
 import school.faang.user_service.mapper.UserMapperImpl;
 import school.faang.user_service.repository.mentorship.MentorshipRepository;
 import school.faang.user_service.exception.DataValidationException;
-import school.faang.user_service.service.user.filter.UserFilter;
+import school.faang.user_service.filters.user.UserFilter;
 
 import java.util.ArrayList;
 import java.util.List;
