+++ src/test/java/school/faang/user_service/service/subscription/SubscriptionServiceTest.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.service.service.subscription;
+package school.faang.user_service.service.subscription;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -13,8 +13,7 @@ import school.faang.user_service.entity.User;
 import school.faang.user_service.exception.DataValidationException;
 import school.faang.user_service.mapper.ShortUserMapper;
 import school.faang.user_service.repository.SubscriptionRepository;
-import school.faang.user_service.service.subscription.SubscriptionService;
-import school.faang.user_service.service.user.filter.UserFilter;
+import school.faang.user_service.filters.user.UserFilter;
 
 import java.util.List;
 import java.util.stream.Stream;
